+++ src/clojure2d/extra/utils.clj
@@ -114,4 +114,3 @@
     (with-canvas-> c
       (image (get-image img)))
     (show-window {:canvas c})))
-