+++ utils/palettes/resources/clean.sh
@@ -1,0 +1,1 @@
+rm -rf gradients palettes cpt-city color_presets.edn cpt-city.zip