+++ utils/palettes/resources/copy.sh
@@ -1,0 +1,3 @@
+cp color_presets.edn ../../../resources
+cp -r gradients ../../../resources
+cp -r palettes ../../../resources