+++ utils/palettes/resources/prepare.sh
@@ -1,6 +1,6 @@
-rm -rf gradients palettes cpt-city
+./clean.sh 
 mkdir -p gradients palettes
-wget --no-hsts -O cpt-city.zip http://soliton.vm.bytemark.co.uk/pub/cpt-city/pkg/cpt-city-c3g-2.22.zip
+wget --no-hsts -O cpt-city.zip http://soliton.vm.bytemark.co.uk/pub/cpt-city/pkg/cpt-city-c3g-2.24.zip
 unzip -q cpt-city.zip
 cd cpt-city
 find . -name "*.c3g" > cpt-city-names.txt